Proof-read count_streamlines
