+++ scripts/scil_count_streamlines.py
@@ -10,18 +10,19 @@ import argparse
 import json
 import os
 
-import nibabel as nib
-
-from scilpy.io.streamlines import lazy_streamlines_count
 from scilpy.io.utils import add_json_args, assert_inputs_exist
+from scilpy.tractograms.tractogram_operations import lazy_streamlines_count
 
 
 def _build_arg_parser():
-    p = argparse.ArgumentParser(
-        description=__doc__,
-        formatter_class=argparse.RawTextHelpFormatter)
+    p = argparse.ArgumentParser(description=__doc__,
+                                formatter_class=argparse.RawTextHelpFormatter)
     p.add_argument('in_tractogram',
                    help='Path of the input tractogram file.')
+    p.add_argument('--print_count_alone', action='store_true',
+                   help="If true, prints the result only. \nElse, prints the "
+                        "bundle name and count formatted as a json dict."
+                        "(default)")
     add_json_args(p)
     return p
 
@@ -33,14 +34,17 @@ def main():
     assert_inputs_exist(parser, args.in_tractogram)
 
     bundle_name, _ = os.path.splitext(os.path.basename(args.in_tractogram))
+    count = int(lazy_streamlines_count(args.in_tractogram))
 
-    stats = {
-        bundle_name: {
-            'streamline_count': int(lazy_streamlines_count(args.in_tractogram))
+    if args.print_count_alone:
+        print(count)
+    else:
+        stats = {
+            bundle_name: {
+                'streamline_count': count
+            }
         }
-    }
-
-    print(json.dumps(stats, indent=args.indent))
+        print(json.dumps(stats, indent=args.indent))
 
 
 if __name__ == '__main__':
